\name{Geocode-package}
\docType{package}
\title{
  \packageTitle{Geocode}
}
\description{
  \packageDescription{Geocode}
}
\details{

  The DESCRIPTION file:
    \packageDESCRIPTION{Geocode}
  \packageIndices{Geocode}
  The lab contains creation of a package which will connect to google geocode API
  and gives us the desired locations full address, latitude and longitude.
}
\author{ packageAuthor {Ashraful}

  Maintainer packageMaintainer{Ashraful}
}
\keyword{ geocode,api,location }
\examples{geocode_api("dhaka")
}